manage folder and move 2 files
